# api/v1/openapi.yaml
...
swagger: '2.0'
info:
  title: Cilium API
  description: Cilium
  version: v1beta
...
basePath: "/v1beta"
produces:
- application/json
consumes:
- application/json
paths:
...
  "/endpoint/{id}/config":
    get:
      summary: Retrieve endpoint configuration
      description: |
        Retrieves the configuration of the specified endpoint.
      tags:
      - endpoint
      parameters:
      - "$ref": "#/parameters/endpoint-id"
      responses:
        '200':
          description: Success
          schema:
            "$ref": "#/definitions/Configuration"
        '404':
          description: Endpoint not found
    patch:
      summary: Modify mutable endpoint configuration
      description: |
        Update the configuration of an existing endpoint and regenerates &
        recompiles the corresponding programs automatically.
      tags:
      - endpoint
      parameters:
      - "$ref": "#/parameters/endpoint-id"
      - name: configuration
        in: body
        required: true
        schema:
          "$ref": "#/definitions/ConfigurationMap"
      responses:
        '200':
          description: Success
        '400':
          description: Invalid configuration request
          x-go-name: Invalid
        '404':
          description: Endpoint not found
        '500':
          description: Update failed. Details in message.
          x-go-name: Failed
          schema:
            "$ref": "#/definitions/Error"
  "/endpoint/{id}/labels":
    get:
      summary: Retrieves the list of labels associated with an endpoint.
      tags:
      - endpoint
      parameters:
      - "$ref": "#/parameters/endpoint-id"
      responses:
        '200':
          description: Success
          schema:
            "$ref": "#/definitions/LabelConfiguration"
        '404':
          description: Endpoint not found
    put:
      summary: Modify label configuration of endpoint
      description: |
        Updates the list of labels associated with an endpoint by applying
        a label modificator structure to the label configuration of an
        endpoint.

        The label configuration mutation is only executed as a whole, i.e.
        if any of the labels to be deleted are not either on the list of
        orchestration system labels, custom labels, or already disabled,
        then the request will fail. Labels to be added which already exist
        on either the orchestration list or custom list will be ignored.
      tags:
      - endpoint
      parameters:
      - "$ref": "#/parameters/endpoint-id"
      - name: configuration
        in: body
        required: true
        schema:
          "$ref": "#/definitions/LabelConfigurationModifier"
      responses:
        '200':
          description: Success
        '404':
          description: Endpoint not found
        '460':
          description: Label to be deleted not found
          x-go-name: LabelNotFound
          schema:
            "$ref": "#/definitions/Error"
        '500':
          description: Error while updating labels
          x-go-name: UpdateFailed
          schema:
            "$ref": "#/definitions/Error"
...
parameters:
  endpoint-id:
    name: id
    description: |
      String describing an endpoint with the format `[prefix:]id`. If no prefix
      is specified, a prefix of `cilium-local:` is assumed. Not all endpoints
      will be addressable by all endpoint ID prefixes with the exception of the
      local Cilium UUID which is assigned to all endpoints.

      Supported endpoint id prefixes:
        - cilium-local: Local Cilium endpoint UUID, e.g. cilium-local:3389595
        - cilium-global: Global Cilium endpoint UUID, e.g. cilium-global:cluster1:nodeX:452343
        - container-id: Container runtime ID, e.g. container-id:22222
        - container-name: Container name, e.g. container-name:foobar
        - pod-name: pod name for this container if K8s is enabled, e.g. pod-name:default:foobar
        - docker-net-endpoint: Docker libnetwork endpoint ID, e.g. docker-net-endpoint:4444
    in: path
    required: true
    type: string
...
definitions:
...
  Labels:
    description: Set of labels
    type: array
    items:
      type: string
  LabelConfiguration:
    description: Label configuration of an endpoint
    type: object
    properties:
      orchestration-identity:
        description: "Labels derived from orchestration system that are used in computing a security identity"
        "$ref": "#/definitions/Labels"
      orchestration-info:
        description: "Labels derived from orchestration system that are not used in computing a security identity"
        "$ref": "#/definitions/Labels"
      custom:
        description: "Custom labels in addition to orchestration system labels."
        "$ref": "#/definitions/Labels"
      disabled:
        description: "Labels derived from orchestration system which have been disabled."
        "$ref": "#/definitions/Labels"
  LabelConfigurationModifier:
    description: |
      Structure describing label mutations to be performed on a
      LabelConfiguration object.
    type: object
    properties:
      add:
        description: |
          List of labels to add and enable. If the label is an orchestration
          system label which has been disabled before, it will be removed from
          the disabled list and readded to the orchestration list. Otherwise
          it will be added to the custom label list.
        "$ref": "#/definitions/Labels"
      delete:
        description: |
          List of labels to delete. If the label is an orchestration system
          label, then it will be deleted from the orchestration list and
          added to the disabled list. Otherwise it will be removed from the
          custom list.
        "$ref": "#/definitions/Labels"
...
  Configuration:
    description: |
      General purpose structure to hold configuration of the daemon and
      endpoints. Split into a mutable and immutable section.
    type: object
    properties:
      mutable:
        description: Changeable configuration
        "$ref": "#/definitions/ConfigurationMap"
      immutable:
        description: Immutable configuration (read-only)
        "$ref": "#/definitions/ConfigurationMap"
      policy-enforcement:
        type: string
  ConfigurationMap:
    description: |
      Map of configuration key/value pairs.
    type: object
    additionalProperties:
      type: string
...
  Error:
    type: string